--allow-incompatible-update: exclude pinned pkgs
Don't throw errors about changing versions of root dependencies from
the previous solution if the package has a top-level equality
constraint like foo@=1.2.3.

This avoids the need for --allow-incompatible-update in a variety of
cases that it isn't really designed for:

* Core package versions changing when switching branches or changing
  between prereleases.
* Scenarios where you override a core package
* Adding an explicit equality constraint: `meteor add foo@=1.2.3`
  (you will still be prompted to pass --allow-incompatible-update
  if the constraint is inexact, like foo@1.2.3).

Fixes the 'list-with-a-new-version' self-test, 'package-depends-on-
either-version', and probably 'old cli tests'.