Remove broken do-not-update-to-rcs test
It was broken because fake-warehouse and test-package-server were
incompatible.  In order to actually use the published packages, we had
to sync test-package-server into the local warehouse.  But as soon as we
did that, the latest release became some other random release.

In any case, the semantics tested here have changed twice since 0.9.0!
In 0.9.0.1 we changed it from "ask for any RC means get any RC" to
"don't get any RCs that aren't explicitly anticipated".  And in 1.1
we're changing it to "minimize unanticipated RCs".

Manually running through the equivalent of this test, most of it works
except for the last line; `meteor update` will not take you from one RC
to the next, due to minimizing unanticipated RCs.  That's probably fine
though.

While we're at it, make it much more clear that warehouse and
test-package-server don't work together.