Improve Logic.variableBits
check that the second argument is present

generate better variable names.  the generated
names don't have to be magic -- we will probably
document what they are -- but they should be a
little better than just appending a number.
it's reasonable to give variables names like
"1,1", and appending a 0 to that would make
"1,10", which is weird.  "1,1$0" is better.