The tool only needs the ddp client
However, turns out the previous commit doesn't really work --- making
webapp a strong dependency of ddp-server means that as soon as we load
mongo we'll load webapp in the tool which is bad.  So move the mongodb
Npm module out of mongo, since that's all we need in the tool.