FIx mongo failover self-test
mongo now transitively uses webapp (because ddp-server does) so we can't
really make an app with mongo and no webapp.  This is a bit of a shame
and when we refactor the mongo package we should fix this... but until
then, let's just remove the no-longer-necessary main function and be OK
with this test running a pointless web server.