Fix test-packages after an initial build failure
The comments immediately after this new code claimed that the "changes
we made were persisted to disk", but they weren't.  So, if you had some
sort of build error that occured after the initializeCatalog stage (say,
invalid JS in a non-package.js file), when the runner reset the context,
it would lose the changes to
projectContext.projectConstraintsFile (.meteor/packages), and you'd end
up with no packages in the test runner app.