Windows installer: Fix on Windows Server 2012
Turns out there was unused code that dealt with strings
that may or may not have been initialized, and treated
them as multi-byte strings even though they were unicode
strings.

I don't actually understand why this would cause a failure
only on Windows Server 2012, but this change definitely fixed
the problem, and the code removed was unnecessary and wrong anyways.

While at it, improved error reporting in case we fail to
get the link to the bootstrap tarball, or download it.

Also fixed a bug that broke login/signup (due to passing a
char* instead of a wchar_t*)