Fix copy-dev-bundle-from-jenkins.sh
awscli doesn't print full paths with slashes, so splitting on slashes is
wrong.  And I'm not sure what `>/dev/null | wc -l` was supposed to
detect, but it didn't work due to the spaces wc prints.

It's possible that there's an aws cli version skew here. We should
rewrite this as a node script just using some node s3 module.