spacebars-compiler: Enforce the {{> React}} no sibling rule
The react-template-helper package exports a {{> React}} template that
can be used to insert React components into Blaze templates. Due to
the interface that `React.render()` provides, React components rendered
this way must be placed as the only child of their parent element.

Prior to this change, we had runtime (as opposed to compile time) tests
in the react-template-helper package. These tests couldn't detect all
error cases (such as when the sibling is a conditional that starts
empty).

It's unfortunate that this change will make all template compilation
slower but we could improve that later with a change @glasser described
where build plugins can know what other build plugins are running for
the app being built.

TODO: Consider adding an option to `Spacebars.codeGen` with the name
of the template being compiled and add that to the error. This would
bring these errors to the same level as the runtime ones currently
thrown in the react-template-helper package.