Use a FiberPool instead of a CallbackQueue.
It didn't make sense for all Promise callbacks to be scheduled serially
with respect to one another, but it's also important that we don't create
a new Fiber for every Promise callback that ever fires.

The FiberPool abstraction is especially nice because we can use it to
implement Promise.async, too.