Use a wrapper NPM package for double-ended-queue.
Most immediately, this techique serves as a workaround for
https://github.com/npm/read-installed/issues/40, since the latest version
of [double-ended-queue](https://www.npmjs.com/package/double-ended-queue)
is still a prerelease version. This issue is going to rear its head the
next time we rebuild node for the dev bundle, and it will be very
difficult to diagnose the problem later if we don't fix it now.

Longer-term, we could move things like `Meteor._SynchronousQueue` into the
meteor-deque package, but that's a discussion for a different commit.