Always use identity.email for the email field for github accounts
Code introduced in f32f07a can error out because identity.email might
be empty and primaryEmail might be undefined at the same time.

Let’s always use the publicly visible email for the email field and if
scope is enough to provide the GitHub account’s list of emails then we
only add that to the emails field and developers which use the user or
user:email scope should use the emails field instead of email field.

If the scope is revoked, the list of emails will become empty in the
emails field.