Reuse previous Constraint Solver result in tool
If the same ProjectContext calls the PackagesResolver
multiple times and the ConstraintSolver.Input structure
is identical, reuse the same answer.  We still ask the
tool for the catalog data, which currently goes to sqlite
(taking ~70ms in the app tested), but we skip creating a
CS.Solver and a Logic.Solver and running them (~450ms).

Note that if a new ProjectContext is created (not sure when
that happens?), there's no caching across that.