fix `meteor update --packages-only` in checkout
If you're running from a checkout but your project has a non-none
release file, the release file should always be ignored. But this one
bit of code (basing the package update code on the post-update release)
paid attention to it anyway.  This tended to lead to a lot of errors
like:

  No version of webapp satisfies all constraints: @=1.2.1, @=1.2.0