Install es5-shim by default for new apps.
The es5-shim package can be removed by expert users who do not wish to
support older browsers, or who wish to support older browsers using a
different shim library.

To enable that use case, we do not register a strong dependency on the
es5-shim package in the core meteor package, but we do ensure that
es5-shim loads first when it is installed.