Remove JSON package
JSON support is in all browsers we could conceivably care about; it's
ES5 and has perfect support in IE 8.  No need to ship a library as
a core package, and no need to specify dependencies on it.

We could conceivably publish a new version of the `json` package that
is empty, so that apps using packages that still say `use("json")`
also get the code size reduction, but we'll wait until someone requests
that.