fix for cordova plugins [not real] conflict
Band-aidy, but fixes the case when two packages add the same plugin with
the same version that can't be parsed by semver (like github url).

For some reason, after the build plugins refactoring the same plugin
from the same package can now be added twice from `webapp`.

Since this feature is not well designed to begin with, decided to spend
too much time investigating it.