Don't try to interact with fences that have fired
This bug was introduced by #4694 which switched OplogObserveDriver's
listener from using beginWrite to the new onBeforeFire.  beginWrite
doesn't throw an error when called on a retired fence; onBeforeFire
does.  So don't try to interact with fired fences. (I'm not sure if
there is an importance to the distinction between retired and fired
introduced by dcd26415f, but this code should be fine.)

While we're at it, make the error in question (which shouldn't happen)
be delivered to Mongo write callbacks (or thrown), if for no other
reason than that it allows us to test this fix.

Fixes #4839.

Conflicts:
	packages/mongo/mongo_livedata_tests.js