Fix HTML.isConstructedObject and comment
The comments are not related to the fix, they are just for future
reference.

The fix is that having an "own" property named `constructor` should
not disqualify an object from being considered an "instance of some
class."  Instead, we should detect that an object like
`{constructor: ...}` is a plain object by noting that
`!(x instanceof x.constructor)`.