Undocumented $METEOR_FORCE_PORTABLE option
If set, it assumes that all local packages are portable and can be
deployed to any platform.

If you are using this, you need to consistently use it every time you
run Meteor, not just when deploying.  The first time you use it in a
project that has been used without it, you should run `meteor rebuild`.