Convert tarball URLs in cordova-plugins in upgrader
Although we convert tarball URLs when building, and we still need to do
this for plugins added through packages, seeing the old URLs in `meteor
list` may confuse people. So we now attempt to convert them as part of
upgrading, together with the plugin ID changes. This also gives us the
opportunity to warn people right away if they’ve been using non-GitHub
tarball URLs.