ReactiveDict: can set absent value to undefined
Setting a ReactiveDict value to `undefined` does not remove it, and
undefined values are migrated.  Therefore, it is odd that setting
'foo' to `undefined` on the empty dict does not produce the dict
`{foo: undefined}`, but setting 'foo' first to another value and *then*
`undefined` does.  So, we fix that.

This means that setting a value to `undefined` blocks `setDefault` on
that key, which makes sense given that setting the value to 123 and then
`undefined` blocks `setDefault` on that key.