Fix ReactiveDict#delete when no deps listeners
Checking that `self.keyValueDeps[key]` exists is important here.

Also remove alignment, remove comment.  (Deleting the key isn't
"clean-up", it's what actually does the work!  Clean-up would be
removing Deps, but that is tricky and for another day.)

A variable named `oldKeys` doesn't make sense if it's just an alias for
`self.keys`.