Avoid invalidation setting absent key to undefined
Fixes a test-breaking aspect of 60aa9f7, which is that while it made
setting an absent value to "undefined" work, doing so would cause an
invalidation that didn't happen before.

While we're at it, clean up the code and make sure that setting an
absent key to 'undefined' does invalidate "all()".

Unit tests in "session" and "reactive-dict" now pass.