Fix Meteor.startup not waiting for deviceready in Cordova app
This fixes a long standing issue where Meteor.startup was not waiting
for the deviceready event (fired when all plugins have been
initialized) in a Cordova app. It also removes a non-functioning
workaround that was based on the belief that this was a Cordova bug,
and cleans up the code somewhat.

The cause of this issue was that we tried to keep track of
awaitingEventsCount, but forgot that both DOMContentLoaded and loaded
events would usually fire. As a result, awaitingEventsCount became 0
before deviceready fired (leading to all sorts of bugs with code
calling into Cordova).

The holdReady() function is based on a similarly named function in
jQuery and generalizes waiting for events in addition to loading
completing.