Fix client hash calculation during initial build for Cordova
Including an empty PUBLIC_SETTINGS object when there are no public
settings fixes a discrepancy between the way we calculated the client
hash for Cordova in CordovaBuilder and in webapp_server.js. This
discrepancy meant a Cordova app would always download a new version the
first time it started up, because the bundled version hash and the
server version hash would be different.