Use readAndWatchFileWithHash when importing modules.
The options passed to linker.fullLink are stringified and hashed to
produce a cacheKey, so I hesitated when it seemed I might need to pass a
WatchSet object via linkerOptions to the ImportScanner. This hesitation
was fruitful, however, because I realized the ImportScanner can do its
work in PackageSourceBatch#_linkJS, which is where the linkerOptions are
determined, which means we don't have to pass nearly so many
ImportScanner-specific options to the linker.

Fixes #5815.