Refactor reset password and enroll account
Was working with `Accounts.sendResetPasswordEmail` and `Accounts.sendEnrollmentEmail` and noticed a comment that the two could be refactored.  I went ahead and abstracted the 99% of the identical code from both and wrapped it in a 'sendEmail' method.  The only real difference between the two (which were a good 50 lines of code each) was the key `resetPassword` and `enrollAccount`.