Better polyfill for Error.captureStackTrace in Meteor.makeErrorType.
The trick of returning an object other than `this` from the constructor
function makes the classes created by Meteor.makeErrorType impossible to
subclass, unless the subclass is very careful to return the same object
from its own constructor.

Fortunately, the only reason we needed an actual Error instance in this
code was to get access to its .stack property, which we can just as easily
borrow from an auxiliary Error object.