Fix linker's in-memory cache to populate from disk
The linker caches files both on disk and in an in-memory LRU cache.
Because it didn't populate the in-memory cache when it hit the disk
cache, and the disk cache persists between tool runs, the in-memory
cache would typically sit empty.

This fix seems to knock 1.8s off of a 12-second rebuild time for one
app.