Speed improvements to Profile
profile.js uses some "dumb" data structures and algorithms that are
N^2 or even N^3.  If you experiment with adding more data to the
profile (e.g. names of files written), you'll find that the time it
takes to generate the profiler report blows up.  This commit
makes the profiler usable again in that case.

* Don't create a new array in isChild(.), and compare elements
  starting at the end (because early elements are more likely
  to be similar)
* Swap order of entryName(.) (cheap) and isLeaf(.) (expensive)
  in leafTotal(.)