Bump 1.3 beta release numbers.
The beta released with this commit (modules-beta.6) included everything on
the release-1.3 branch up to bdf64da3391b85217048706c7fc752a18624715e
("Avoid circular package.json resolution chains.").

Unfortunately, additional commits were pushed to the branch before this
commit was pushed, so the release tag release/METEOR@1.3-modules-beta.6
does not correspond to any commit in the branch history.

Rather than attempting to rewrite the branch history, I decided to amend
this commit with this explanation before pushing it.

To see exactly what was released with the sixth beta:

  git fetch --tags
  git log release/METEOR@1.3-modules-beta.6