Create noRetry option on Connection.apply
Closes #6108
Pull request #6180

Test is included, as is updated documentation.
All tests in ddp-client package (including new test for #6108) pass.

Implemented as simply a noRetry flag on apply which triggers it to fail with a
Meteor.Error('invocation-failed') instead of retrying on reconnect.

Test passes, methods which are marked as noRetry get an error in case of reset,
only if they've sent at least once, and this happens before messageSent flag is
cleared.  This results in it getting it's callback as soon as quiescence
triggers, and the callback is called with an Error 'invocation-failed'.

[stubailo: changed error code, fixed docs, squashed commits]