files.copyFile: don't stat again if have mode
we were already passing the mode of `from` into `files.copyFile`
in at least one place, but `files.copyFile` was dropping it on
the floor.  now we use it avoid doing a second stat.

Also, list copyFile in the profiler report.  copyFile does its own
custom file I/O, so it didn't show up as a readFile or writeFile
which are automatically profiled.