Ensure every file the ImportScanner scans has a valid sourcePath.
Packages published before the modules package existed may acquire a
dependency on the modules package upon upgrading to Meteor 1.3, because
they depended on the ecmascript package, which now implies the modules
package. If the previously published isopacket does not provide a
sourcePath or path property, then we should use the servePath property
instead. With that fix in place, I am now strictly enforcing that all
files scanned by the ImportScanner actually exist on disk.

Should fix #5871.