`buildMode === "test"` for both `test` and `test-packages`
I think we hadn't done this for back-compat reasons, however it doesn't make sense as it means that test-drivers and test utilities can't be `testOnly` if they want to be usable for `meteor test-packages`. Which kind of makes it pointless to have `testOnly`.

I don't anticipate it will cause problems, but I'm not entirely sure.