Remove duplicate call to `DDPCommon.stringifyDDP`
This code simplification in ddp-server removes the duplicate call to
`DDPCommon.stringifyDDP` when sending the `connected` message.

The change makes the code easier to understand: it clarifies that
there's nothing magical about the `connnected` message: it is sent in
the same way as the other DDP messages.

It is almost a pure refactoring: if the undocumented `_printSentDDP`
option is false (the default), there's no change in the behavior of
the code.

Previously, if `Meteor._printSentDDP` was set, the server would
display other sent messages but not the `connected` message.  With
this change, when the option is set all sent DDP messages will be
displayed.