Fix for `check` max-recursion when called on circular references
This is a stop-gap (???) fix for a bigger issue with `EJSON.stringify`
working on circular objects.  #4778 *apparently* fixes most *other*
`EJSON` methods but it's not even clear it if it would fix the
`.stringify` problem properly (after all, `JSON.stringify` doesn't
either).  Also, that PR is over a year old.

This fix takes the approach of trying to fix the calling of
`EJSON.stringify` with an object that it can't work with.  It's doing
this by trying the native `JSON.stringify` to see if it would work.  If
that returns a `TypeError`, then it just dumbs-down the object to
the `typeof`.

Also added a lot of tests

I can't say it's the absolute best, but it *will* take care of:
* Fix #6712
* Fix #4587

Depends on #6736 (Match.Optional working properly)